session.send() echoes to the textlog by default echo on new line if last line doesn't end with a > (prompt) added #sql command to query the world db added Plugin.send() method which defaults to an orange1 color for the echo Don't echo account name, password
