+++ abacura-core/abacura/plugins/debug/pycharm.py
@@ -18,4 +18,4 @@ class PycharmDebug(Plugin):
             self.output(f"Connecting to {host}:{port}")
             self.debugger.connect(host, port)
         except ConnectionRefusedError:
-            self.session.output(f"Connection refused by pycharm debugger {host}:{port}")
+            self.output(f"Connection refused by pycharm debugger {host}:{port}")
